+++ charts/alluxio/CHANGELOG.md
@@ -197,6 +197,7 @@
 0.8.0
 
 - Change worker from Daemonset to Statefulset
+- Support container network
 
 0.9.0